requirements.txt
opencv-python
ultralytics
onnxruntime
pillow
supervision
imread_from_url
onnx